refactoring - moved every recognizer related component to common directory, rewrote the Readme file
