+++ index.html
@@ -5,7 +5,7 @@
     <link rel="icon" href="/favicon.ico"/>
     <link rel="stylesheet" href="https://use.typekit.net/cpg2hdr.css">
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <title>Learn to Speak</title>
+    <title>Learn To Speak</title>
 </head>
 <body>
 <div id="app"></div>
