+++ src/components/recognizer/ExerciseSlider.vue
