+++ src/components/recognizer/RecognizerCounter.vue
