+++ src/components/recognizer/RecognizerHeader.vue
@@ -15,7 +15,7 @@
 <script lang="ts">
 import {defineComponent} from "vue";
 import {useCounterStore} from "@/stores/counter";
-import RecognizerCounter from "@/components/RecognizerCounter.vue";
+import RecognizerCounter from "@/components/recognizer/RecognizerCounter.vue";
 
 export default defineComponent({
   name: 'RecognizerHeader',
