+++ src/router/index.ts
@@ -1,5 +1,5 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import Recognizer from "@/components/Recognizer.vue";
+import Recognizer from "@/components/recognizer/Recognizer.vue";
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
